% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\docType{data}
\name{collaboration}
\alias{collaboration}
\title{Collaboration networks}
\format{A tibble with 1722 rows and 3 columns:
\describe{
\item{id_interview}{ID number of the interview}
\item{from,to}{Node IDs unique within the interview of two collaborating scientists}
}}
\source{
\url{https://recon.icm.edu.pl}
}
\usage{
collaboration
}
\description{
Edge list of an undirected graph linking scientists with collaboration ties.
}
\keyword{datasets}
