% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconqdata-package.R
\docType{package}
\name{reconqdata-package}
\alias{reconqdata}
\alias{reconqdata-package}
\title{reconqdata: RECON project: Qualitative Data}
\description{
Egocentric network data from a qualitative study on collaboration in science conducted in the RECON project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://recon.icm.edu.pl}
}

}
\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (0000-0001-7503-852X)

Authors:
\itemize{
  \item Dominika Czerniawska
  \item Wojciech Fenrich
}

Other contributors:
\itemize{
  \item National Science Centre [funder]
}

}
\keyword{internal}
